@@ -1,6 +1,7 @@
+'use client'
 import React from 'react'
 import { Button } from '../ui/button'
-import { useRouter } from 'next/router'
+import { useRouter } from 'next/navigation'
 
 const CreateNootbook = () => {
   const router = useRouter()
