@@ -1,15 +1,24 @@
 import React from 'react'
+import { getNotes } from '@/services/nootbook.service'
+import Slug from '@/components/tiptap/slug'
 
 interface Slug {
   params: {
     slug: string
   }
 }
-const Page = ({ params }: Slug) => {
+const Page = async ({ params }: Slug) => {
   console.log(params)
+  const content = await getNotes(params.slug)
+  console.log(!content.success)
+  if (!content.success) {
+    return <div>
+      <p>NotFound</p>
+    </div>
+  }
   return (
     <div>
-      <p>{params.slug}</p>
+      <Slug content={content} />
     </div>
   )
 }
